Struct molecules::application::Application

pub struct Application {
    width: u32,
    height: u32,
    device: Device,
    queue: Queue,
    start_time: SystemTime,
    pub camera: RotationCamera,
    pub camera_changed: bool,
    voxel_grid: VoxelGrid,
    raymarch_globals: RaymarchGlobals,
    raymarch_globals_buffer: Buffer,
    ssao_globals_buffer: Buffer,
    raymarch_pipeline: RaymarchPipeline,
    render_pipeline: RenderPipeline,
    ssao_pipeline: SsaoPipeline,
    gbuffer_positions: TextureView,
    gbuffer_normals: TextureView,
    output_texture: TextureView,
    sdf_default: Buffer,
    sdf_texture: Texture,
    sdf_texture_view: TextureView,
    mouse_pressed: bool,
    mouse_position: PhysicalPosition<f64>,
}

Fields
width: u32

Width of the window

height: u32

Height of the window

device: Device

Device used for rendering

queue: Queue

Main queue to which GPU commands are sent

start_time: SystemTime

Time of initilization of program. Used for animation.

camera: RotationCamera

Camera of the application.

camera_changed: bool

Holds information whether camera was changed between frames. The information is used for accumulation of result.

voxel_grid: VoxelGrid

Voxel grid containing atoms of the molecule.

raymarch_globals: RaymarchGlobals

Global variables for ray marching passed to GPU.

raymarch_globals_buffer: Buffer

GPU buffer for raymarch_globals.

ssao_globals_buffer: Buffer

Global variables for SSAO computation passed to GPU.

raymarch_pipeline: RaymarchPipeline

Pipeline for ray marching.

render_pipeline: RenderPipeline

Pipeline that renders the sphere marched result to the window.

ssao_pipeline: SsaoPipeline

Pipeline that adds SSAO to the sphere marched result.

gbuffer_positions: TextureViewgbuffer_normals: TextureViewoutput_texture: TextureViewsdf_default: Buffersdf_texture: Texture

Texture where signed distance field is stored. Used to progressively enhance view when camera did not change between frames.

sdf_texture_view: TextureViewmouse_pressed: boolmouse_position: PhysicalPosition<f64>
Methods

impl Application

pub async fn new<'_, '_>(
width: u32,
height: u32,
surface: &'_ Surface,
file_name: Option<&'_ str>
) -> Self

Initialized the application.

pub fn resize(&mut self, width: u32, height: u32)

Called when window is resized. Recreates textures for rendering.

pub fn render(&mut self, frame: &TextureView)

Called each frame to render.

pub fn window_event(&mut self, event: &WindowEvent)

pub fn device_event(&mut self, event: &DeviceEvent)

pub fn device(&self) -> &Device

Returns reference to the device used by the application.

pub fn queue_mut(&self) -> &Queue

Returns reference to the device used by the application.

fn update_raymarch_globals(&mut self)

pub fn solvent_radius(&self) -> f32

pub fn set_solvent_radius(&mut self, solvent_radius: f32)

pub fn max_neighbours(&self) -> i32

pub fn set_max_neighbours(&mut self, max_neighbours: i32)